separated the old API to a submodule (#163)
* separated the old API to a submodule

* switched off codecov for unused files
