+++ wannierberri/__old_API/__init__.py
@@ -1,0 +1,5 @@
+""" This submodule collects old functions, that may not be supported in the future,
+when their  functionality is complitely moved to newer APIs 
+The code outside of this submodule should not depend on this code,
+except for explicit calls to API `integrate()` and `tabulate()`
+"""
