+++ wannierberri/__old_API/__integrate.py
@@ -15,9 +15,9 @@ import numpy as np
 from collections import defaultdict
 from copy import copy
 
-from .__utility import TAU_UNIT
-from .smoother import VoidSmoother
-from . import __result as result
+from wannierberri.smoother import VoidSmoother
+from wannierberri.__factors import TAU_UNIT
+from wannierberri import __result as result
 from . import fermiocean as fermiocean
 from . import __kubo as kubo
