+++ wannierberri/__result_tab.py
@@ -1,142 +1,10 @@
-#                                                            #l
-# This file is distributed as part of the WannierBerri code  #
-# under the terms of the GNU General Public License. See the #
-# file `LICENSE' in the root directory of the WannierBerri   #
-# distribution, or http://www.gnu.org/copyleft/gpl.txt       #
-#                                                            #
-# The WannierBerri code is hosted on GitHub:                 #
-# https://github.com/stepan-tsirkin/wannier-berri            #
-#                     written by                             #
-#           Stepan Tsirkin, University of Zurich             #
-#                                                            #
-#------------------------------------------------------------
-
 import numpy as np
-from collections.abc import Iterable
-from collections import defaultdict
 from time import time
 import multiprocessing
 import functools
+import itertools
+from collections.abc import Iterable
 from . import __result as result
-from . import covariant_formulak as frml
-from . import covariant_formulak_basic as frml_basic
-
-#If one whants to add  new quantities to tabulate, just modify the following dictionaries
-
-#should be classes Formula_ln
-# TODO : add factors to the calculation
-calculators = {
-    'spin': frml.Spin,
-    'V': frml.Velocity,
-    'berry': frml.Omega,
-    'Der_berry': frml.DerOmega,
-    'morb': frml.morb,
-    'Der_morb': frml_basic.Der_morb,
-    'spin_berry': frml.SpinOmega,
-}
-
-additional_parameters = defaultdict(lambda: defaultdict(lambda: None))
-additional_parameters_description = defaultdict(lambda: defaultdict(lambda: "no description"))
-
-descriptions = defaultdict(lambda: "no description")
-descriptions['berry'] = "Berry curvature (Ang^{2})"
-descriptions['Der_berry'] = "1st deravetive of Berry curvature (Ang^{3})"
-descriptions['V'] = "velocity (eV*Ang)"
-descriptions['spin'] = "Spin"
-descriptions['morb'] = "orbital moment of Bloch states <nabla_k u_n| X(H-E_n) | nabla_k u_n> (eV*Ang**2)"
-descriptions[
-    'Der_morb'] = "1st derivative orbital moment of Bloch states <nabla_k u_n| X(H-E_n) | nabla_k u_n> (eV*Ang**2)"
-descriptions['spin_berry'] = "Spin Berry curvature (hbar * Ang^{2})"
-
-parameters_ocean = {
-    'external_terms': (True, "evaluate external terms"),
-    'internal_terms': (True, "evaluate internal terms"),
-}
-
-for key, val in parameters_ocean.items():
-    for calc in ['berry', 'Der_berry', 'morb', 'Der_morb']:
-        additional_parameters[calc][key] = val[0]
-        additional_parameters_description[calc][key] = val[1]
-
-additional_parameters['spin_berry']['spin_current_type'] = 'ryoo'
-additional_parameters_description['spin_berry']['spin_current_type'] = 'type of the spin current'
-
-
-def tabXnk(
-        data_K,
-        quantities=[],
-        user_quantities={},
-        degen_thresh=-1,
-        degen_Kramers=False,
-        ibands=None,
-        parameters={},
-        specific_parameters={}):
-
-    if ibands is None:
-        ibands = np.arange(data_K.nbands)
-    else:
-        ibands = np.array(ibands)
-
-    tabulator = Tabulator(data_K, ibands, degen_thresh=degen_thresh, degen_Kramers=degen_Kramers)
-
-    results = {'Energy': result.KBandResult(data_K.E_K[:, ibands], TRodd=False, Iodd=False)}
-    for qfull in quantities:
-        q = qfull.split('^')[0]
-        __parameters = {}
-        for param in additional_parameters[q]:
-            if param in parameters:
-                __parameters[param] = parameters[param]
-            else:
-                __parameters[param] = additional_parameters[q][param]
-        results[qfull] = tabulator(calculators[q](data_K, **__parameters))
-
-    for q, formula in user_quantities.items():
-        if q in specific_parameters:
-            __parameters = specific_parameters[q]
-        else:
-            __parameters = {}
-        results[q] = tabulator(formula(data_K, **__parameters))
-
-    return TABresult(kpoints=data_K.kpoints_all, recip_lattice=data_K.system.recip_lattice, results=results)
-
-
-class Tabulator():
-
-    def __init__(self, data_K, ibands, degen_thresh=1e-4, degen_Kramers=False):
-
-        self.nk = data_K.nk
-        self.NB = data_K.num_wann
-        self.ibands = ibands
-
-        band_groups = data_K.get_bands_in_range_groups(
-            -np.Inf, np.Inf, degen_thresh=degen_thresh, degen_Kramers=degen_Kramers, sea=False)
-        # bands_groups  is a digtionary (ib1,ib2):E
-        # now select only the needed groups
-        self.band_groups = [
-            [n for n in groups.keys() if np.any((ibands >= n[0]) * (ibands < n[1]))] for groups in band_groups
-        ]  # select only the needed groups
-        self.group = [[] for ik in range(self.nk)]
-        for ik in range(self.nk):
-            for ib in self.ibands:
-                for n in self.band_groups[ik]:
-                    if ib < n[1] and ib >= n[0]:
-                        self.group[ik].append(n)
-                        break
-
-    def __call__(self, formula):
-        """formula  - TraceFormula to evaluate"""
-        rslt = np.zeros((self.nk, len(self.ibands)) + (3, ) * formula.ndim)
-        for ik in range(self.nk):
-            values = {}
-            for n in self.band_groups[ik]:
-                inn = np.arange(n[0], n[1])
-                out = np.concatenate((np.arange(0, n[0]), np.arange(n[1], self.NB)))
-                values[n] = formula.trace(ik, inn, out) / (n[1] - n[0])
-            for ib, b in enumerate(self.ibands):
-                rslt[ik, ib] = values[self.group[ik][ib]]
-
-        return result.KBandResult(rslt, TRodd=formula.TRodd, Iodd=formula.Iodd)
-
 
 class TABresult(result.Result):
 
@@ -414,6 +282,29 @@ class TABresult(result.Result):
     def max(self):
         return -1  # tabulating does not contribute to adaptive refinement
 
+def write_frmsf(frmsf_name, Ef0, numproc, quantities, res, suffix=""):
+    if len(suffix) > 0:
+        suffix = "-" + suffix
+    if frmsf_name is not None:
+        open(f"{frmsf_name}_E{suffix}.frmsf", "w").write(res.fermiSurfer(quantity=None, efermi=Ef0, npar=numproc))
+        ttxt = 0
+        twrite = 0
+        for Q in quantities:
+            for comp in res.results[Q].get_component_list():
+                try:
+                    t31 = time()
+                    txt = res.fermiSurfer(quantity=Q, component=comp, efermi=Ef0, npar=numproc)
+                    t32 = time()
+                    open(f"{frmsf_name}_{Q}-{comp}{suffix}.frmsf", "w").write(txt)
+                    t33 = time()
+                    ttxt += t32 - t31
+                    twrite += t33 - t32
+                except result.result.NoComponentError:
+                    pass
+    else:
+        ttxt = 0
+        twrite = 0
+    return ttxt, twrite
 
 def _savetxt(limits=None, a=None, fmt=".8f", npar=0):
     assert a.ndim == 1, "only 1D arrays are supported. found shape{}".format(a.shape)
@@ -437,26 +328,122 @@ def _savetxt(limits=None, a=None, fmt=".8f", npar=0):
         return "".join(res)
 
 
-def write_frmsf(frmsf_name, Ef0, numproc, quantities, res, suffix=""):
-    if len(suffix) > 0:
-        suffix = "-" + suffix
-    if frmsf_name is not None:
-        open(f"{frmsf_name}_E{suffix}.frmsf", "w").write(res.fermiSurfer(quantity=None, efermi=Ef0, npar=numproc))
-        ttxt = 0
-        twrite = 0
-        for Q in quantities:
-            for comp in res.results[Q].get_component_list():
+class KBandResult(result.Result):
+
+    def __init__(self, data, TRodd, Iodd):
+        if isinstance(data, list):
+            self.data_list = data
+        else:
+            self.data_list = [data]
+        self.TRodd = TRodd
+        self.Iodd = Iodd
+
+    def fit(self, other):
+        for var in ['TRodd', 'Iodd', 'rank', 'nband']:
+            if getattr(self, var) != getattr(other, var):
+                return False
+        return True
+
+    @property
+    def data(self):
+        if len(self.data_list) > 1:
+            self.data_list = [np.vstack(self.data_list)]
+        return self.data_list[0]
+
+    @property
+    def rank(self):
+        return len(self.data_list[0].shape) - 2
+
+    @property
+    def nband(self):
+        return self.data_list[0].shape[1]
+
+    @property
+    def nk(self):
+        return sum(data.shape[0] for data in self.data_list)
+
+    def __add__(self, other):
+        assert self.fit(other)
+        return KBandResult(self.data_list + other.data_list, self.TRodd, self.Iodd)
+
+    def __mul__(self, number):
+        return KBandResult([d * number for d in self.data_list], self.TRodd, self.Iodd)
+
+    def __truediv__(self, number):
+        return self * 1  # actually a copy
+
+    def to_grid(self, k_map):
+        dataall = self.data
+        data = np.array([sum(dataall[ik] for ik in km) / len(km) for km in k_map])
+        return KBandResult(data, self.TRodd, self.Iodd)
+
+    def select_bands(self, ibands):
+        return KBandResult(self.data[:, ibands], self.TRodd, self.Iodd)
+
+    def average_deg(self, deg):
+        for i, D in enumerate(deg):
+            for ib1, ib2 in D:
+                for j in range(len(self.data_list)):
+                    self.data_list[j][i, ib1:ib2] = self.data_list[j][i, ib1:ib2].mean(axis=0)
+        return self
+
+    def transform(self, sym):
+        data = [sym.transform_tensor(data, rank=self.rank, TRodd=self.TRodd, Iodd=self.Iodd) for data in self.data_list]
+        return KBandResult(data, self.TRodd, self.Iodd)
+
+    def get_component_list(self):
+        dim = len(self.data.shape[2:])
+        return ["".join(s) for s in itertools.product(*[("x", "y", "z")] * dim)]
+
+    def get_component(self, component=None):
+        xyz = {"x": 0, "y": 1, "z": 2}
+        dims = np.array(self.data.shape[2:])
+        if not np.all(dims == 3):
+            raise RuntimeError(f"dimensions of all components should be 3, found {dims}")
+
+        ndim = len(dims)
+
+        if component is not None:
+            component = component.lower()
+        if component == "":
+            component = None
+        if ndim == 0:
+            if component is None:
+                return self.data
+            else:
+                raise result.NoComponentError(component, 0)
+        elif ndim == 1:
+            if component in ["x", "y", "z"]:
+                return self.data[:, :, xyz[component]]
+            elif component == 'norm':
+                return np.linalg.norm(self.data, axis=-1)
+            elif component == 'sq':
+                return np.linalg.norm(self.data, axis=-1)**2
+            else:
+                raise result.NoComponentError(component, 1)
+        elif ndim == 2:
+            if component == "trace":
+                return sum([self.data[:, :, i, i] for i in range(3)])
+            else:
                 try:
-                    t31 = time()
-                    txt = res.fermiSurfer(quantity=Q, component=comp, efermi=Ef0, npar=numproc)
-                    t32 = time()
-                    open(f"{frmsf_name}_{Q}-{comp}{suffix}.frmsf", "w").write(txt)
-                    t33 = time()
-                    ttxt += t32 - t31
-                    twrite += t33 - t32
-                except result.NoComponentError:
-                    pass
-    else:
-        ttxt = 0
-        twrite = 0
-    return ttxt, twrite
+                    return self.data[:, :, xyz[component[0]], xyz[component[1]]]
+                except IndexError:
+                    raise result.NoComponentError(component, 2)
+        elif ndim == 3:
+            if component == "trace":
+                return sum([self.data[:, :, i, i, i] for i in range(3)])
+            else:
+                try:
+                    return self.data[:, :, xyz[component[0]], xyz[component[1]], xyz[component[2]]]
+                except IndexError:
+                    raise result.NoComponentError(component, 3)
+        elif ndim == 4:
+            if component == "trace":
+                return sum([self.data[:, :, i, i, i, i] for i in range(3)])
+            else:
+                try:
+                    return self.data[:, :, xyz[component[0]], xyz[component[1]], xyz[component[2]], xyz[component[3]]]
+                except IndexError:
+                    raise result.NoComponentError(component, 4)
+        else:
+            raise NotImplementedError("writing tensors with rank >4 is not implemented. But easy to do")
